% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_features}
\alias{time_features}
\title{time features example: IBM and Microsoft Close Prices}
\format{
A data frame with 2 columns and 1324 rows.
}
\source{
finance.yahoo.com
}
\usage{
time_features
}
\description{
A data frame with with daily with daily prices for IBM and Microsoft since March 2017.
}
\keyword{datasets}
